fix: mv getHeaders from context.ts to workflow-requests.ts
didn't want getHeaders in the public API.
